Adds sidebar and basic functionality
Adds sidebar as a nav element with a UL for the menu items. Current
styling lacks color for the selected tab and mobile view icons are not
centered vertically. I play to use the table cell trick to get them to
move to the proper location. Links are in place so that once I figure
out how to target the items from the query parameter, I should be set to
go.