Adds highlight JS along with styling
Found jQuery code to replace <img> tags with a specific class with
actual <svg> objects allowing for highlight to change the fill of them.
Once the page loads, it creates an array of the different parameters and
another function uses the value for tag to add a class to the
corresponding menu item. The class then adds the Sense red color.